+++ block-streamer/src/bitmap_processor.rs
