+++ block-streamer/src/main.rs
@@ -1,7 +1,7 @@
 use tracing_subscriber::prelude::*;
 
 mod bitmap;
-mod block_height_stream;
+mod bitmap_processor;
 mod block_stream;
 mod delta_lake_client;
 mod graphql;
